+++ src/Tests/AppGateway.fs
@@ -9,6 +9,7 @@ open Farmer.ApplicationGateway
 open Farmer.Builders
 open Farmer.Network
 open Farmer.NetworkSecurity
+open Newtonsoft.Json.Linq
 
 let client =
     new NetworkManagementClient(Uri "http://management.azure.com", TokenCredentials "NotNullOrWhiteSpace")
@@ -157,8 +158,6 @@ let tests =
                 add_resources [ msi; net; myNsg; myAppGateway ]
             }
 
-            deployment.Template |> Writer.toJson |> ignore
-
             let resource =
                 deployment
                 |> findAzureResources<Microsoft.Azure.Management.Network.Models.ApplicationGateway>
@@ -223,6 +222,7 @@ let tests =
             Expect.hasLength resource.RequestRoutingRules 1 "Expecting 1 request routing rule"
             let routingRule = resource.RequestRoutingRules.[0]
             Expect.equal routingRule.Name "web-front-to-services-back" "Routing rule name did not match"
+            Expect.equal routingRule.Priority (Nullable(1000)) "Routing rule must have a priority"
 
             Expect.equal
                 routingRule.HttpListener.Id
@@ -239,4 +239,163 @@ let tests =
                 "[resourceId('Microsoft.Network/applicationGateways/backendHttpSettingsCollection', 'app-gw', 'bp-default-web-80-web-80')]"
                 "Routing rule http settings did not match"
         }
+
+        test "HTTPS App Gateway" {
+            let myNsg = nsg {
+                name "agw-nsg"
+
+                add_rules [
+                    securityRule {
+                        name "app-gw"
+                        description "GatewayManager"
+                        services [ NetworkService("GatewayManager", Range(65200us, 65535us)) ]
+                        add_source_tag TCP "GatewayManager"
+                        add_destination_any
+                    }
+                    securityRule {
+                        name "inet-gw"
+                        description "Internet to gateway"
+                        services [ "https", 443 ]
+                        add_source_tag TCP "Internet"
+                        add_destination_network "10.28.0.0/24"
+                    }
+                    securityRule {
+                        name "app-servers"
+                        description "Internal app server access"
+                        services [ "http", 80 ]
+                        add_source_network TCP "10.28.0.0/24"
+                        add_destination_network "10.28.1.0/24"
+                    }
+                ]
+            }
+
+            let net = vnet {
+                name "agw-vnet"
+
+                build_address_spaces [
+                    addressSpace {
+                        space "10.28.0.0/16"
+
+                        subnets [
+                            subnetSpec {
+                                name "gw"
+                                size 24
+                                network_security_group myNsg
+                            }
+                            subnetSpec {
+                                name "apps"
+                                size 24
+                                network_security_group myNsg
+                            }
+                        ]
+                    }
+                ]
+            }
+
+            let msi = createUserAssignedIdentity "agw-msi"
+
+            let backendPoolName = ResourceName "agw-be-pool"
+
+            let myAppGateway =
+                let gwIp = gatewayIp {
+                    name "app-gw-ip"
+                    link_to_subnet net.Name net.Subnets.[0].Name
+                }
+
+                let frontendIp = frontendIp {
+                    name "app-gw-fe-ip"
+                    public_ip "agp-gw-pip"
+                }
+
+                let frontendPort = frontendPort {
+                    name "port-443"
+                    port 443
+                }
+
+                let listener = httpListener {
+                    name "https-listener"
+                    protocol Protocol.Https
+                    frontend_ip frontendIp
+                    frontend_port frontendPort
+                    backend_pool backendPoolName.Value
+                    ssl_certificate "ag-test-cert"
+                }
+
+                let backendPool = appGatewayBackendAddressPool {
+                    name backendPoolName.Value
+                    add_backend_addresses [ backend_ip_address "10.28.1.4"; backend_ip_address "10.28.1.5" ]
+                }
+
+                let healthProbe = appGatewayProbe {
+                    name "agw-probe"
+                    host "localhost"
+                    path "/"
+                    port 80
+                    protocol Protocol.Http
+                }
+
+                let backendSettings = backendHttpSettings {
+                    name "bp-default-web-80-web-80"
+                    port 80
+                    probe healthProbe
+                    protocol Protocol.Http
+                    request_timeout 10<Seconds>
+                }
+
+                let routingRule = basicRequestRoutingRule {
+                    name "web-front-to-services-back"
+                    http_listener listener
+                    backend_address_pool backendPool
+                    backend_http_settings backendSettings
+                }
+
+                appGateway {
+                    name "app-gw"
+                    sku_capacity 2
+                    add_identity msi
+                    add_ip_configs [ gwIp ]
+                    add_frontends [ frontendIp ]
+                    add_frontend_ports [ frontendPort ]
+                    add_http_listeners [ listener ]
+                    add_backend_address_pools [ backendPool ]
+                    add_backend_http_settings_collection [ backendSettings ]
+                    add_request_routing_rules [ routingRule ]
+                    add_probes [ healthProbe ]
+
+                    add_ssl_certificates [
+                        sslCertificate {
+                            name "ag-test-cert"
+                            key_vault_secret_id "https://my-kv.vault.azure.net/secrets/app-gw-cert"
+                        }
+                    ]
+
+                    depends_on myNsg
+                    depends_on net
+                }
+
+            let deployment = arm {
+                location Location.EastUS
+                add_resources [ msi; net; myNsg; myAppGateway ]
+            }
+
+            let jobj = deployment.Template |> Writer.toJson |> JObject.Parse
+            let appGwProps = jobj.SelectToken("resources[?(@.name=='app-gw')].properties")
+
+            let appGwHttpsCertId =
+                appGwProps.SelectToken("httpListeners[0].properties.sslCertificate.id")
+
+            Expect.equal
+                (appGwHttpsCertId |> string)
+                "[resourceId('Microsoft.Network/applicationGateways/sslCertificates', 'app-gw', 'ag-test-cert')]"
+                "Generated incorrect references to HTTPS cert"
+
+            let appGwSslCert = appGwProps.SelectToken("sslCertificates[0]")
+            Expect.equal (appGwSslCert["name"] |> string) "ag-test-cert" "Wrong name on SSL Certificate"
+
+            Expect.equal
+                (appGwSslCert.SelectToken("properties.keyVaultSecretId") |> string)
+                "https://my-kv.vault.azure.net/secrets/app-gw-cert"
+                "Wrong Key Vault Secret ID for SSL Certificate"
+        }
+
     ]